refactor: pack fix bottom logic into custom hook
